# openapi.yaml
openapi: 3.1.0
info:
    title: 'Marketplace API'
    version: 1.0.0
    description: 'Multi-vendor marketplace API — Express, TypeScript, Prisma. Auto-generated from Zod schemas.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
security:
    -
        bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
...
paths:
...
    /stores:
        get:
            tags:
                - catalogue
            parameters:
                -
                    schema:
                        type: string
                    required: false
                    name: q
                    in: query
                -
                    schema:
                        type: string
                        pattern: ^\d+$
                    required: false
                    name: page
                    in: query
                -
                    schema:
                        type: string
                        pattern: ^\d+$
                    required: false
                    name: limit
                    in: query
            responses:
                '200':
                    description: 'List of stores'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    items: {type: array, items: {type: object, properties: {id: {type: string}, ownerId: {type: string}, name: {type: string}, slug: {type: string}, logoUrl: {type: [string, 'null']}, owner: {type: object, properties: {id: {type: string}, email: {type: string}}, required: [id, email]}}, required: [id, ownerId, name, slug, logoUrl]}}
                                    meta: {type: object, properties: {total: {type: integer}, page: {type: integer}, pageSize: {type: integer}}, required: [total, page, pageSize]}
                                required:
                                    - items
                                    - meta
        post:
            tags:
                - catalogue
            security:
                -
                    bearerAuth: []
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    minLength: 1
                                slug:
                                    type: string
                                    pattern: '^[a-z0-9-]+$'
                                logoUrl:
                                    type: string
                                    format: uri
                            required:
                                - name
                                - slug
            responses:
                '201':
                    description: 'Store created successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id: {type: string}
                                    ownerId: {type: string}
                                    name: {type: string}
                                    slug: {type: string}
                                    logoUrl: {type: [string, 'null']}
                                    owner: {type: object, properties: {id: {type: string}, email: {type: string}}, required: [id, email]}
                                required:
                                    - id
                                    - ownerId
                                    - name
                                    - slug
                                    - logoUrl
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '409':
                    description: 'You already own a store or store slug already exists'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
    '/stores/{slug}':
        get:
            tags:
                - catalogue
            parameters:
                -
                    schema:
                        type: string
                        description: 'Resource slug'
                    required: true
                    description: 'Resource slug'
                    name: slug
                    in: path
            responses:
                '200':
                    description: 'Store details'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id: {type: string}
                                    ownerId: {type: string}
                                    name: {type: string}
                                    slug: {type: string}
                                    logoUrl: {type: [string, 'null']}
                                    owner: {type: object, properties: {id: {type: string}, email: {type: string}}, required: [id, email]}
                                required:
                                    - id
                                    - ownerId
                                    - name
                                    - slug
                                    - logoUrl
                '404':
                    description: 'Store not found'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
        put:
            tags:
                - catalogue
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        description: 'Resource slug'
                    required: true
                    description: 'Resource slug'
                    name: slug
                    in: path
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    minLength: 1
                                slug:
                                    type: string
                                    pattern: '^[a-z0-9-]+$'
                                logoUrl:
                                    type: string
                                    format: uri
            responses:
                '200':
                    description: 'Store updated successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id: {type: string}
                                    ownerId: {type: string}
                                    name: {type: string}
                                    slug: {type: string}
                                    logoUrl: {type: [string, 'null']}
                                    owner: {type: object, properties: {id: {type: string}, email: {type: string}}, required: [id, email]}
                                required:
                                    - id
                                    - ownerId
                                    - name
                                    - slug
                                    - logoUrl
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '403':
                    description: 'Access denied - you do not own this store'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '404':
                    description: 'Store not found or access denied'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '409':
                    description: 'Store slug already exists'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
        delete:
            tags:
                - catalogue
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        description: 'Resource slug'
                    required: true
                    description: 'Resource slug'
                    name: slug
                    in: path
            responses:
                '204':
                    description: 'Store deleted successfully'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '403':
                    description: 'Access denied - you do not own this store'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                '404':
                    description: 'Store not found or access denied'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
...